move playback functions down to playback controls level
