+++ src/components/PlaybackControls.tsx
@@ -1,28 +1,17 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { Button, InputNumber, Slider, Tooltip } from 'antd';
 
 import '../../css/playback_controls.css';
-import { PlaybackState } from '../constants';
-import { TimeUnits } from '../constants';
+import { PlaybackData, PlaybackState } from '../constants';
 import { FrameBack, FrameForward, Pause, Play } from './Icons';
+import { SimulariumController, compareTimes } from '@aics/simularium-viewer';
 
 const TOOLTIP_COLOR = '#3B3649';
 interface PlayBackProps {
   playbackState: PlaybackState;
-  timeUnits: TimeUnits;
-  firstFrameTime: number;
-  lastFrameTime: number;
-  timeStep: number;
-  displayTimes: DisplayTimes;
-  playHandler: () => void;
-  pauseHandler: () => void;
-  prevHandler: () => void;
-  nextHandler: () => void;
-  handleSliderChange: (time: number | [number, number]) => void;
-  handleSliderAfterChange: (time: any) => void;
-  handleTimeInputChange: (event: any) => void;
-  handleTimeInputKeyDown: (event: any) => void;
-  scaleBarLabel: string;
+  setPlaybackState: (playbackState: PlaybackState) => void;
+  playbackData: PlaybackData;
+  controller: SimulariumController;
 }
 
 interface DisplayTimes {
@@ -32,28 +21,54 @@ interface DisplayTimes {
   roundedTimeStep: number;
 }
 
-const PlayBackControls = (props: PlayBackProps): JSX.Element => {
-  const {
-    playbackState,
-    timeUnits,
-    firstFrameTime,
-    lastFrameTime,
-    timeStep,
-    displayTimes,
-    nextHandler,
-    prevHandler,
-    handleSliderChange,
-    handleSliderAfterChange,
-    playHandler,
-    pauseHandler,
-    handleTimeInputChange,
-    handleTimeInputKeyDown,
-  } = props;
+const roundTimeForDisplay = (time: number) => {
+  if (time === 0) {
+    return 0;
+  }
 
+  return parseFloat(time.toPrecision(3));
+};
+
+const getDisplayTimes = (
+  playbackState: PlaybackState,
+  playbackData: PlaybackData
+): DisplayTimes => {
+  const { currentTime: currentTime } = playbackState;
+  const { timeUnits, firstFrameTime, lastFrameTime, timeStep } = playbackData;
+  return {
+    roundedTime: roundTimeForDisplay(currentTime * timeUnits.magnitude),
+    roundedFirstFrameTime: roundTimeForDisplay(
+      firstFrameTime * timeUnits.magnitude
+    ),
+    roundedLastFrameTime: roundTimeForDisplay(
+      lastFrameTime * timeUnits.magnitude
+    ),
+    roundedTimeStep: roundTimeForDisplay(timeStep * timeUnits.magnitude),
+  };
+};
+
+const PlayBackControls = (props: PlayBackProps): JSX.Element => {
+  const { playbackState, setPlaybackState, playbackData, controller } = props;
+  const [timeInput, setTimeInput] = useState(0);
+
+  const { timeUnits, firstFrameTime, lastFrameTime, timeStep } = playbackData;
   const { isPlaying, currentTime } = playbackState;
 
   const isStepForwardDisabled = currentTime + timeStep >= lastFrameTime;
   const isStepBackDisabled = firstFrameTime + timeStep > currentTime;
+
+  const displayTimes = getDisplayTimes(playbackState, playbackData);
+
+  // use effect listener on current time, if current time is end, call reset playback
+  useEffect(() => {
+    if (currentTime + timeStep >= lastFrameTime) {
+      controller.gotoTime(firstFrameTime);
+      setPlaybackState({
+        currentTime: firstFrameTime,
+        isPlaying: false,
+      });
+    }
+  }, [currentTime]);
 
   // Determine the width of the input box based on a likely max number of characters
   const getTimeInputWidth = (): string => {
@@ -68,13 +83,105 @@ const PlayBackControls = (props: PlayBackProps): JSX.Element => {
     }ch`;
   };
 
+  const playHandler = (timeOverride?: number) => {
+    const { currentTime } = playbackState;
+    const newTime = timeOverride !== undefined ? timeOverride : currentTime;
+    const timeIsBeyondLastFrameTime = newTime + timeStep >= lastFrameTime;
+    controller.pause();
+    if (timeIsBeyondLastFrameTime) {
+      setPlaybackState({
+        isPlaying: false,
+        currentTime: firstFrameTime,
+      });
+      controller.gotoTime(firstFrameTime);
+    } else {
+      setPlaybackState({
+        isPlaying: true,
+        currentTime: newTime,
+      });
+      controller.playFromTime(newTime);
+      controller.resume();
+    }
+  };
+
+  const pauseHandler = () => {
+    controller.pause();
+    setPlaybackState({ ...playbackState, isPlaying: false });
+  };
+
+  const gotoNextFrame = () => {
+    const { currentTime } = playbackState;
+    controller.gotoTime(currentTime + timeStep);
+  };
+
+  const goToPreviousFrame = () => {
+    const { currentTime } = playbackState;
+    controller.gotoTime(currentTime - timeStep);
+  };
+
+  const handleTimeInputChange = (userInput: number | null): void => {
+    if (userInput !== null) {
+      setTimeInput(userInput as number);
+    }
+  };
+
+  const skipToTime = (time: number) => {
+    const isTimeGreaterThanLastFrameTime =
+      compareTimes(time, lastFrameTime, timeStep) === 1;
+    const isTimeLessThanFirstFrameTime =
+      compareTimes(time, firstFrameTime, timeStep) === -1;
+    if (isTimeGreaterThanLastFrameTime || isTimeLessThanFirstFrameTime) {
+      return;
+    }
+    controller.gotoTime(time);
+  };
+
+  const handleTimeInputKeyDown = (
+    event: React.KeyboardEvent<HTMLInputElement>
+  ): void => {
+    if (event.key === 'Enter' || event.key === 'Tab') {
+      // User input will be aligned with the displayed time values, which were multiplied
+      // by timeUnits.magnitude in the getDisplayTimes selector, so we have to undo the
+      // multiplication before requesting the time. timeUnits.magnitude is 1 for a vast
+      // majority of the time so it shouldn't make a difference most times.
+      if (typeof timeInput === 'number') {
+        setPlaybackState({
+          ...playbackState,
+          currentTime: timeInput / timeUnits.magnitude,
+        });
+        skipToTime(timeInput / timeUnits.magnitude);
+      }
+    }
+    if (event.key === 'Escape') {
+      const inputNumberComponent = event.target as HTMLElement;
+      inputNumberComponent.blur();
+    }
+  };
+
+  const handleSliderChange = (sliderValue: number | [number, number]): void => {
+    const { isPlaying } = playbackState;
+    skipToTime(sliderValue as number);
+    if (isPlaying) {
+      controller.pause();
+    }
+  };
+
+  const handleSliderAfterChange = (value: number): void => {
+    const { isPlaying } = playbackState;
+    if (isPlaying) {
+      playHandler(value);
+    } else {
+      setPlaybackState({ ...playbackState, currentTime: value });
+    }
+  };
+
   return (
     <div className="playback-controls">
       <Tooltip placement="top" title="Skip 1 frame back" color={TOOLTIP_COLOR}>
         <Button
           id={'back-button'}
           className="btn"
-          onClick={prevHandler}
+          onClick={goToPreviousFrame}
           disabled={isStepBackDisabled}
         >
           {FrameBack}
@@ -103,7 +210,7 @@ const PlayBackControls = (props: PlayBackProps): JSX.Element => {
         <Button
           id={'next-button'}
           className="btn"
-          onClick={nextHandler}
+          onClick={gotoNextFrame}
           disabled={isStepForwardDisabled}
         >
           {FrameForward}
